<3.10-Updating from PgConnection to Pool (According to my understanding) The connection pool to the server required by the sqlx requires a mutable reference to the connection. It's discussed that it is possible to use a Mutex for such purpose, however possible, not ideal because a slow connection could end up blocking all the process completely. The solution was to changing from a Connection to a Pool of connections. Under the hood, providing a pool, would allow the SQLX to manage the connections.
The reason for it to need a mut reference is because the SQLX wants
to make sure that there won't be another connection working at the same
time that SQLX is processing the connection, which it achieves by making
taking advantage of the single &mut borrow of the compile checker.

What other big things we did?
-We expand on the subscription code to not just return an ok, but by
also accesing the Database to insert data.
-We updated from a random port to a specific port.
-Add a copy of the database pool to the HTTP app data. (web::Data<T>
is basically an Arc<T>)