<3.11 -*Summary
Because I'm using the "antonio" as user for the db, I had an issue with connecting to a DB to create DBs.

The solution was to create a "dummy database" which in this case I called "antonio" so I could run my tests with. Also, in my todo for the next commit is to implement deleting the database created during tests because....... I'm too lazy to reset the image everytime it gets too cramped lol