=*4.5.13-Protect Your Secrets - secrecy
People might say it's boilerplate to do this, however, the fact that everything will throw error once you wrap the sensitive data is amazing because the compiler will be stuck with "What ** is this?" during compilation, and force you to rethink what you're exposing. And if there's ever an accidental "leak", you only have to look for the "expose_secret()" method.