*5.3.8 Image Optimization - Scraping /target
By docker-ignoring all the irrelevant files, we can end up with a ~100MB image for the Docker.

Worth noting is that the Entrypoints are the binary name.

Also, I couldn't figure how to manage how to tap into the health checkpoint from the host machine. I'll later upload an update on what changes were made to the project to be able to tap into the checkpoint when I can grab some time to read into someone's guide for the Docker Compose method.