6.5 Ownership Meets Invariants
The gist of this chapter, is how do we ensure that data inserted into the database through "insert_subscriber" is correct? locally speaking, it's impossible to verify that the data is correct (i.e. empty name) without checking it... We could check it, but then we would waste resources to check data that's fine. Hence there comes the need to accept a data-type that we can be confident that's correct without needing to perform aditional checks, hence the need to refactor into another data type, shifting from a `"form: &UserEmail"` to `"new_subscriber: &NewSubscriber"`